pyproject.toml
[build-system]
requires = [
	"setuptools",
        "Mathics3-Module-Base",
        "Mathics3>9.0.0",
        "PatternLite",
        "langid",  # replace with a supported newer package, e.g. via spacy
        "matplotlib",
        "mpmath>=1.2.0",
        "numpy",
        "pycountry>=3.2.0",
        "pyenchant>=3.2.0",
        "scipy>=1.10.0",
        "spacy>=3.4",
        "wasabi<1.1.0,>=0.8.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "Mathics3-Module-nltk"
description = "Mathics3 Natural Language Toolkit module"
dependencies = [
        "Mathics3>=9.0.0",
        "Mathics3-Module-Base>=9.0.0",
        "click>=8.0",
        "joblib>=1.0.1",
        "langid",  # replace with a supported newer package, e.g. via spacy
        "llvmlite>=0.36",
        "nltk>=3.8.0",
        "mpmath>=1.2.0",
        "PatternLite",
        "pyenchant>=3.2.0",
        "pycountry>=3.2.0",
        "spacy>=3.4",
        "wasabi<1.1.0,>=0.8.2",
]
requires-python = ">=3.10"
readme = "README.rst"
license = "GPL-3.0-or-later"
keywords = ["Mathematica", "Wolfram", "Interpreter", "Shell", "Math", "CAS"]
maintainers = [
    {name = "Mathics Group", email = "mathics-devel@googlegroups.com"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Interpreters",
]

dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/Mathics3/Mathics3-Module-nltk"
Downloads = "https://github.com/Mathics3/Mathics3-Module-nltk/releases"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools]
packages = [
    "pymathics.natlang",
]

[tool.setuptools.dynamic]
version = {attr = "pymathics.natlang.version.__version__"}
